+++ auth_user_service/db_models/__init__.py
@@ -5,3 +5,4 @@ from .sessions import ClientSession as ClientSession
 from .api_keys import ApiKey as ApiKey, RateLimit as RateLimit
 from .tombstones import AuthTombstone as AuthTombstone
 from .security_policy import SecurityPolicy as SecurityPolicy
+from .outbox import RevocationOutbox as RevocationOutbox